backend/src/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-ai-formrecognizer==3.3.3
azure-ai-inference==1.0.0b9
azure-ai-translation-text==1.0.1
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
fastapi==0.115.12
h11==0.16.0
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.2.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
